Refactor avoid_final_with_getter_rule.dart and avoid_final_with_getter_visitor.dart
